requirements.txt
flask
flask-cors
python-dotenv
sqlalchemy
alembic
psycopg2-binary
pyjwt
bcrypt
google-generativeai
logging
pyyaml
jwt_auth
requests
pydantic